requirements.txt
asgiref==3.7.2
bleach==6.1.0
Django==3.2.24
django-summernote==0.8.20.0
django-taggit==5.0.1
mysql-connector-python==8.3.0
mysqlclient==1.4.6
pillow==10.2.0
PyMySQL==1.1.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
webencodings==0.5.1